git-svn-id: http://192.168.0.243:4435/svn/bee/branches/transrifus@2318 aaee961a-53e7-2249-bd01-b2a8f0c4ce08

+++ src/com/butent/bee/client/modules/ec/widget/ItemList.java
@@ -258,7 +258,7 @@ public class ItemList extends Flow implements KeyDownHandler {
     return panel;
   }
 
-  private static Widget renderStock(int stock, String unit) {
+  private static Widget renderStock(double stock, String unit) {
     Flow wrapper = new Flow(STYLE_STOCK_WRAPPER);
 
     Widget stockWidget = EcWidgetFactory.createStockWidget(stock);
